// @types/streamData.d.ts
// Generated by https://quicktype.io

export interface StreamData {
	items:                    Item[];
	watchHistorySaveInterval: number;
	errors?:                  Error[]
}

export interface Error {
	detail: string,
	code: number
}

export interface Item {
	src:          string;
	kind:         string;
	isPromo:      boolean;
	videoType:    string;
	aips:         Aip[];
	experienceId: string;
	showAds:      boolean;
	id:           number;
}

export interface Aip {
	out: number;
	in:  number;
}